+++ lib/app/data/models/money_objects/investments/stock_cumulative.dart
@@ -3,5 +3,6 @@ import 'package:money/app/core/helpers/ranges.dart';
 class StockCumulative {
   double amount = 0.00;
   DateRange dateRange = DateRange();
+  double dividend = 0.00;
   double quantity = 0.0;
 }