+++ lib/app/data/models/money_objects/investments/stock_cumulative.dart
@@ -3,6 +3,14 @@ import 'package:money/app/core/helpers/ranges.dart';
 class StockCumulative {
   double amount = 0.00;
   DateRange dateRange = DateRange();
-  double dividend = 0.00;
+  List<Dividend> dividends = [];
+  double dividendsSum = 0.00;
   double quantity = 0.0;
+}
+
+class Dividend {
+  Dividend(this.date, this.amount);
+
+  final double amount;
+  final DateTime date;
 }